requirements.txt
fastapi
psycopg2-binary 
uvicorn